fit_xy {generics}	R Documentation

Estimate model parameters.

Description

Estimates parameters for a given model from a set of data in the form of a set of predictors (x) and outcome(s) (y).

Usage

fit_xy(object, ...)

Arguments

object
An object. See the individual method for specifics.

...
Other arguments passed to methods

Methods

No methods found in currently loaded packages.

[Package generics version 0.1.2 Index]