interpolate {generics}	R Documentation

Interpolate missing values

Description

Interpolates missing values provided in the training dataset using the fitted model.

Usage

interpolate(object, ...)

Arguments

object
A fitted model object

...
Other arguments passed to methods

Value

A dataset (tibble::tibble() or similar) of the same structure as the input dataset with missing values from the response variable replaced with interpolated values.

Methods

No methods found in currently loaded packages.

[Package generics version 0.1.2 Index]