tune_args {generics}	R Documentation

Determine arguments tagged for tuning

Description

tune_args() takes an object such as a model specification or a recipe and returns a tibble of information on all possible tunable arguments and whether or not they are actually tunable.

Usage

tune_args(object, ...)

Arguments

object
A model_spec, recipe, workflow, or other object.

...
Other arguments passed to methods.

Details

The source column is determined differently for a model_spec or a recipe (with additional detail on the type).

The id field has any identifier that was passed from tune::tune() (e.g. tune("some note")). If no additional detail was used in that function, the id field reverts to the name of the parameters.

Value

A tibble with columns for the parameter name (name), whether it contains any tunable value (tune), the id for the parameter (id), and the information on where the parameter was located (source).

Methods

No methods found in currently loaded packages.

[Package generics version 0.1.3 Index]