components {generics}	R Documentation

Extract components

Description

components can be used to extract elements from an object.

Usage

components(object, ...)

Arguments

object
A data separable object.

...
Other arguments passed to methods

Details

For example, decomposition methods and some modelling techniques can be used to decompose a dataset into components of interest. This function is used to extract these components in a tidy data format.

Value

A dataset (tibble::tibble() or similar) containing components from the object.

Methods

No methods found in currently loaded packages.

[Package generics version 0.1.2 Index]