min_grid {generics}	R Documentation

Determine the minimum set of model fits

Description

min_grid() determines exactly what models should be fit in order to evaluate the entire set of tuning parameter combinations. This is for internal use only and the API may change in the near future.

Usage

min_grid(x, grid, ...)

Arguments

x
A model specification.

grid
A tibble with tuning parameter combinations.

...
Not currently used.

Value

A tibble with the minimum tuning parameters to fit and an additional list column with the parameter combinations used for prediction.

Methods

No methods found in currently loaded packages.

[Package generics version 0.1.2 Index]